fix: Renamed manager file to keep name convention
