+++ src/database/__init__.py
@@ -1,3 +1,3 @@
-from .db_object_manager import DatabaseObjectManager
+from .database_manager import DatabaseObjectManager
 from .database_writer import Writer
 from .database_reader import  Reader
